support LocalDateType and UtilDateType
